+++ includes/ray.h
@@ -6,11 +6,11 @@
 
 typedef struct s_hit
 {
-	float			t;
+	double			t;
 	t_coordinate	hit_point;
 	t_color			hit_color;
 	t_vec			hit_normal;
-	float			min;
+	double			min;
 	int				in_object;	// 0이면 물체 밖에 카메라, 1이면 물체 안에 카메라
 	int				ray_hit;	// 0이면 이 ray는 hit이 없음. 1이면 이 ray는 hit한 ray
 }	t_hit;
@@ -21,7 +21,7 @@ typedef struct s_ray
 	t_hit	hit;
 }	t_ray;
 
-t_ray	init_ray(float x, float y, float z);
+t_ray	init_ray(double x, double y, double z);
 void	shoot_ray(t_mlx *m, t_scene *scene);
 // void	trace_objects(t_mlx *m, t_scene *scene);
